backend/app/api/auth.py
from fastapi import APIRouter, HTTPException, Depends
from pydantic import BaseModel
from datetime import datetime, timedelta
from jose import jwt, JWTError
from typing import Optional

from app.core.config import settings

router = APIRouter()

class AuthRequest(BaseModel):
    message: str
    signature: str
    address: str

class AuthResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"
    expires_in: int

@router.post("/signin", response_model=AuthResponse)
async def signin_with_ethereum(auth_data: AuthRequest):
    """
    Sign in with Ethereum (SIWE) implementation
    TODO: Implement actual signature verification
    """
    # For now, create a simple JWT token
    # In production, verify the signature against the message and address
    
    token_data = {
        "sub": auth_data.address.lower(),
        "iat": datetime.utcnow(),
        "exp": datetime.utcnow() + timedelta(hours=settings.JWT_EXPIRATION_HOURS)
    }
    
    access_token = jwt.encode(
        token_data, 
        settings.JWT_SECRET, 
        algorithm=settings.JWT_ALGORITHM
    )
    
    return AuthResponse(
        access_token=access_token,
        expires_in=settings.JWT_EXPIRATION_HOURS * 3600
    )

@router.get("/me")
async def get_current_user():
    """Get current authenticated user info"""
    # TODO: Implement JWT verification and user retrieval
    return {"address": "0x742d...3456", "authenticated": True}